add chat realtime

+++ Audivia.API/Audivia.API/Controllers/ChatRoomMember/ChatRoomMembersController.cs
@@ -1,6 +1,8 @@
-﻿using Audivia.Application.Services.Interface;
+﻿using Audivia.API.Hubs;
+using Audivia.Application.Services.Interface;
 using Audivia.Domain.ModelRequests.ChatRoomMember;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.SignalR;
 
 namespace Audivia.API.Controllers.ChatRoomMember
 {
@@ -9,16 +11,19 @@ namespace Audivia.API.Controllers.ChatRoomMember
     public class ChatRoomMembersController : ControllerBase
     {
         private readonly IChatRoomMemberService _chatRoomMemberService;
-
-        public ChatRoomMembersController(IChatRoomMemberService chatRoomMemberService)
+        private readonly IHubContext<ChatHub> _hubContext;
+        public ChatRoomMembersController(IChatRoomMemberService chatRoomMemberService, IHubContext<ChatHub> hubContext)
         {
             _chatRoomMemberService = chatRoomMemberService;
+            _hubContext = hubContext;
         }
 
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateChatRoomMemberRequest request)
         {
             var result = await _chatRoomMemberService.CreateChatRoomMember(request);
+            await _hubContext.Clients.Group(request.ChatRoomId)
+            .SendAsync("UserJoined", request);
             return Ok(result);
         }
 
@@ -40,13 +45,21 @@ namespace Audivia.API.Controllers.ChatRoomMember
         public async Task<IActionResult> Update(string id, [FromBody] UpdateChatRoomMemberRequest request)
         {
             await _chatRoomMemberService.UpdateChatRoomMember(id, request);
+            await _hubContext.Clients.Group(id).SendAsync("UserUpdated", request);
             return NoContent();
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(string id)
         {
+            var member = await _chatRoomMemberService.GetMemberById(id);
+            if (member is null)
+            {
+                return NotFound();
+            }
             await _chatRoomMemberService.DeleteChatRoomMember(id);
+            await _hubContext.Clients.Group(member.ChatRoomId).SendAsync("UserLeft", member
+                );
             return NoContent();
         }
     }
